# backend/internal/handler/venueApplication.go
package handler

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// submitVenueApplicationHandler handles requests by venue managers to register a venue
// @Summary      Submit venue application (old)
// @Description  Submit a venue application for approval
// @Tags         manager
// @Accept       json
// @Produce      json
// @Success      200  {object}  map[string]string
// @Router       /api/v1/venues/applications [post]
// @Security     BearerAuth
func (s *Server) submitVenueApplicationHandler(c *gin.Context) {
	// TODO: Implement the venue application submission logic
	c.JSON(http.StatusOK, gin.H{
		"message": "submitVenueApplicationHandler not implemented yet",
	})
}

// approveVenueHandler handles admin approval of a venue
// @Summary      Approve venue
// @Description  Approve a venue application (Admin only)
// @Tags         admin
// @Accept       json
// @Produce      json
// @Param        venue_id path string true "Venue ID"
// @Param        request body map[string]string true "Notes"
// @Success      200  {object}  map[string]any
// @Router       /api/v1/admin/venues/{venue_id}/approve [patch]
// @Security     BearerAuth
func (s *Server) approveVenueHandler(c *gin.Context) {
	venueID := c.Param("venue_id")
	c.JSON(http.StatusOK, gin.H{
		"venue_id":          venueID,
		"onboarding_status": "APPROVED",
		"reviewed_by":       "admin-uuid-123456",
		"admin_notes":       "Approved after reviewing document certifications.",
		"updated_at":        "2026-06-19T20:43:00Z",
	})
}

// rejectVenueHandler handles admin rejection of a venue
// @Summary      Reject venue
// @Description  Reject a venue application (Admin only)
// @Tags         admin
// @Accept       json
// @Produce      json
// @Param        venue_id path string true "Venue ID"
// @Param        request body map[string]string true "Notes"
// @Success      200  {object}  map[string]any
// @Router       /api/v1/admin/venues/{venue_id}/reject [patch]
// @Security     BearerAuth
func (s *Server) rejectVenueHandler(c *gin.Context) {
	venueID := c.Param("venue_id")
	c.JSON(http.StatusOK, gin.H{
		"venue_id":          venueID,
		"onboarding_status": "REJECTED",
		"reviewed_by":       "admin-uuid-123456",
		"admin_notes":       "Rejected: Missing fire safety certificate.",
		"updated_at":        "2026-06-19T20:43:00Z",
	})
}

// listVenuesHandler handles requests to search/list approved venues
// @Summary      List approved venues
// @Description  Get a list of approved venues with optional filters
// @Tags         public-venues
// @Accept       json
// @Produce      json
// @Param        state query string false "State filter"
// @Param        district query string false "District filter"
// @Param        city query string false "City filter"
// @Param        venue_type query string false "Venue type filter"
// @Param        is_air_conditioned query boolean false "Is air conditioned filter"
// @Param        min_seating_capacity query integer false "Min seating capacity filter"
// @Param        limit query integer false "Limit filter"
// @Param        offset query integer false "Offset filter"
// @Success      200  {array}  map[string]any
// @Router       /api/v1/venues [get]
func (s *Server) listVenuesHandler(c *gin.Context) {
	c.JSON(http.StatusOK, []gin.H{
		{
			"venue_id":           "123e4567-e89b-12d3-a456-426614174000",
			"venue_name":         "Grand Palace",
			"city":               "New York",
			"district":           "Manhattan",
			"state":              "New York",
			"venue_type":         "Auditorium",
			"seating_capacity":   500,
			"is_air_conditioned": true,
			"opening_period":     "09:00",
			"closing_period":     "22:00",
		},
	})
}

// getVenueByIDHandler handles requests to view full venue details
// @Summary      Get venue details by ID
// @Description  Get full specifications of an approved venue by its ID
// @Tags         public-venues
// @Accept       json
// @Produce      json
// @Param        venue_id path string true "Venue ID"
// @Success      200  {object}  map[string]any
// @Router       /api/v1/venues/{venue_id} [get]
func (s *Server) getVenueByIDHandler(c *gin.Context) {
	venueID := c.Param("venue_id")
	c.JSON(http.StatusOK, gin.H{
		"venue_id":             venueID,
		"venue_name":           "Grand Palace",
		"address_line_1":       "123 Broadway",
		"address_line_2":       "Suite 100",
		"phone":                "+1234567890",
		"email":                "info@grandpalace.com",
		"city":                 "New York",
		"district":             "Manhattan",
		"state":                "New York",
		"postal_code":          "10001",
		"country_code":         "US",
		"seating_capacity":     500,
		"min_booking_duration": "2h",
		"opening_period":       "09:00",
		"closing_period":       "22:00",
		"relaxation_period":    "30m",
		"is_air_conditioned":   true,
		"venue_type":           "Auditorium",
		"onboarding_status":    "APPROVED",
	})
}

// createManagerVenueHandler handles venue creation by a manager
// @Summary      Create manager venue
// @Description  Create a new venue application (Manager only)
// @Tags         manager
// @Accept       json
// @Produce      json
// @Param        request body map[string]any true "Venue registration body"
// @Success      201  {object}  map[string]any
// @Router       /api/v1/manager/venues [post]
// @Security     BearerAuth
func (s *Server) createManagerVenueHandler(c *gin.Context) {
	c.JSON(http.StatusCreated, gin.H{
		"venue_id":          "123e4567-e89b-12d3-a456-426614174000",
		"onboarding_status": "PENDING_APPROVAL",
	})
}

// listManagerVenuesHandler lists venues owned/managed by the manager
// @Summary      List manager venues
// @Description  Get list of venues owned/managed by the current manager
// @Tags         manager
// @Accept       json
// @Produce      json
// @Success      200  {array}  map[string]any
// @Router       /api/v1/manager/venues [get]
// @Security     BearerAuth
func (s *Server) listManagerVenuesHandler(c *gin.Context) {
	c.JSON(http.StatusOK, []gin.H{
		{
			"venue_id":          "123e4567-e89b-12d3-a456-426614174000",
			"venue_name":         "Grand Palace",
			"city":               "New York",
			"onboarding_status": "PENDING_APPROVAL",
		},
	})
}

// getManagerVenueByIDHandler gets manager's venue by ID
// @Summary      Get manager venue details
// @Description  Get full specifications of manager's own venue by its ID
// @Tags         manager
// @Accept       json
// @Produce      json
// @Param        venue_id path string true "Venue ID"
// @Success      200  {object}  map[string]any
// @Router       /api/v1/manager/venues/{venue_id} [get]
// @Security     BearerAuth
func (s *Server) getManagerVenueByIDHandler(c *gin.Context) {
	venueID := c.Param("venue_id")
	c.JSON(http.StatusOK, gin.H{
		"venue_id":             venueID,
		"venue_name":           "Grand Palace",
		"address_line_1":       "123 Broadway",
		"address_line_2":       "Suite 100",
		"phone":                "+1234567890",
		"email":                "info@grandpalace.com",
		"city":                 "New York",
		"district":             "Manhattan",
		"state":                "New York",
		"postal_code":          "10001",
		"country_code":         "US",
		"seating_capacity":     500,
		"min_booking_duration": "2h",
		"opening_period":       "09:00",
		"closing_period":       "22:00",
		"relaxation_period":    "30m",
		"is_air_conditioned":   true,
		"venue_type":           "Auditorium",
		"onboarding_status":    "PENDING_APPROVAL",
	})
}

// updateManagerVenueHandler updates a manager's venue details
// @Summary      Update manager venue
// @Description  Update details of a manager's venue
// @Tags         manager
// @Accept       json
// @Produce      json
// @Param        venue_id path string true "Venue ID"
// @Param        request body map[string]any true "Venue update body"
// @Success      200  {object}  map[string]any
// @Router       /api/v1/manager/venues/{venue_id} [patch]
// @Security     BearerAuth
func (s *Server) updateManagerVenueHandler(c *gin.Context) {
	venueID := c.Param("venue_id")
	c.JSON(http.StatusOK, gin.H{
		"venue_id":             venueID,
		"venue_name":           "Grand Palace Updated",
		"address_line_1":       "123 Broadway",
		"address_line_2":       "Suite 100",
		"phone":                "+1234567890",
		"email":                "info@grandpalace.com",
		"city":                 "New York",
		"district":             "Manhattan",
		"state":                "New York",
		"postal_code":          "10001",
		"country_code":         "US",
		"seating_capacity":     600,
		"min_booking_duration": "2h",
		"opening_period":       "09:00",
		"closing_period":       "22:00",
		"relaxation_period":    "30m",
		"is_air_conditioned":   true,
		"venue_type":           "Auditorium",
		"onboarding_status":    "PENDING_APPROVAL",
	})
}

// getManagerVenuePricingHandler gets the pricing list of a manager's venue
// @Summary      Get venue pricing list
// @Description  Get pricing tiers and schedules of a manager's venue
// @Tags         manager
// @Accept       json
// @Produce      json
// @Param        venue_id path string true "Venue ID"
// @Success      200  {array}  map[string]any
// @Router       /api/v1/manager/venues/{venue_id}/pricing [get]
// @Security     BearerAuth
func (s *Server) getManagerVenuePricingHandler(c *gin.Context) {
	c.JSON(http.StatusOK, []gin.H{
		{
			"id":             "987f6543-e21b-12d3-a456-426614174000",
			"price_per_hour": 150.00,
			"is_weekend":     false,
			"currency":       "USD",
			"is_active":      true,
			"start_date":     "2026-06-19",
			"end_date":       "2027-06-19",
		},
	})
}

// createManagerVenuePricingHandler creates new pricing for a manager's venue
// @Summary      Create venue pricing
// @Description  Create a new pricing tier for a manager's venue
// @Tags         manager
// @Accept       json
// @Produce      json
// @Param        venue_id path string true "Venue ID"
// @Param        request body map[string]any true "Pricing data"
// @Success      201  {object}  map[string]any
// @Router       /api/v1/manager/venues/{venue_id}/pricing [post]
// @Security     BearerAuth
func (s *Server) createManagerVenuePricingHandler(c *gin.Context) {
	venueID := c.Param("venue_id")
	c.JSON(http.StatusCreated, gin.H{
		"id":             "987f6543-e21b-12d3-a456-426614174000",
		"venue_id":       venueID,
		"price_per_hour": 200.00,
		"is_weekend":     true,
		"currency":       "USD",
		"start_date":     "2026-06-20",
	})
}

// updateManagerVenuePricingHandler updates pricing details of a manager's venue
// @Summary      Update venue pricing
// @Description  Update an existing pricing tier of a manager's venue
// @Tags         manager
// @Accept       json
// @Produce      json
// @Param        venue_id path string true "Venue ID"
// @Param        pricing_id path string true "Pricing ID"
// @Param        request body map[string]any true "Pricing update data"
// @Success      200  {object}  map[string]any
// @Router       /api/v1/manager/venues/{venue_id}/pricing/{pricing_id} [patch]
// @Security     BearerAuth
func (s *Server) updateManagerVenuePricingHandler(c *gin.Context) {
	pricingID := c.Param("pricing_id")
	c.JSON(http.StatusOK, gin.H{
		"id":             pricingID,
		"price_per_hour": 250.00,
		"start_date":     "2026-06-25",
	})
}

// listAdminVenuesHandler lists venues for admin review
// @Summary      List admin venues
// @Description  Get a list of all venues for admin review with optional filters
// @Tags         admin
// @Accept       json
// @Produce      json
// @Param        state query string false "State filter"
// @Param        district query string false "District filter"
// @Param        onboarding_status query string false "Status filter"
// @Param        owner_id query string false "Owner filter"
// @Param        limit query integer false "Limit filter"
// @Param        offset query integer false "Offset filter"
// @Success      200  {array}  map[string]any
// @Router       /api/v1/admin/venues [get]
// @Security     BearerAuth
func (s *Server) listAdminVenuesHandler(c *gin.Context) {
	c.JSON(http.StatusOK, []gin.H{
		{
			"venue_id":          "123e4567-e89b-12d3-a456-426614174000",
			"venue_name":         "Grand Palace",
			"owner_id":          "789e4567-e89b-12d3-a456-426614174000",
			"onboarding_status": "PENDING_APPROVAL",
			"city":               "New York",
			"district":           "Manhattan",
			"state":              "New York",
			"created_at":         "2026-06-19T20:42:12Z",
		},
	})
}

// createBookingHandler creates a new venue booking
// @Summary      Create booking
// @Description  Book a venue with required date and notes
// @Tags         bookings
// @Accept       json
// @Produce      json
// @Param        Idempotency-Key header string true "Idempotency key"
// @Param        request body map[string]any true "Booking parameters"
// @Success      201  {object}  map[string]any
// @Router       /api/v1/bookings [post]
// @Security     BearerAuth
func (s *Server) createBookingHandler(c *gin.Context) {
	c.JSON(http.StatusCreated, gin.H{
		"booking_id":  "456e89b2-12d3-a456-4266-141740001234",
		"venue_id":    "123e4567-e89b-12d3-a456-426614174000",
		"status":      "CONFIRMED",
		"start_time":  "2026-06-25T10:00:00Z",
		"end_time":    "2026-06-25T14:00:00Z",
		"total_price": 600.00,
		"created_at":  "2026-06-19T20:44:12Z",
	})
}

// listBookingsHandler retrieves the list of bookings for the user
// @Summary      List bookings
// @Description  Get list of bookings of the authenticated user
// @Tags         bookings
// @Accept       json
// @Produce      json
// @Param        status query string false "Status filter"
// @Param        limit query integer false "Limit"
// @Param        offset query integer false "Offset"
// @Success      200  {array}  map[string]any
// @Router       /api/v1/bookings [get]
// @Security     BearerAuth
func (s *Server) listBookingsHandler(c *gin.Context) {
	c.JSON(http.StatusOK, []gin.H{
		{
			"booking_id":  "456e89b2-12d3-a456-4266-141740001234",
			"venue_id":    "123e4567-e89b-12d3-a456-426614174000",
			"venue_name":  "Grand Palace",
			"status":      "CONFIRMED",
			"start_time":  "2026-06-25T10:00:00Z",
			"end_time":    "2026-06-25T14:00:00Z",
			"total_price": 600.00,
		},
	})
}

// getBookingByIDHandler retrieves details of a specific booking
// @Summary      Get booking by ID
// @Description  Get details of a specific booking by its ID
// @Tags         bookings
// @Accept       json
// @Produce      json
// @Param        booking_id path string true "Booking ID"
// @Success      200  {object}  map[string]any
// @Router       /api/v1/bookings/{booking_id} [get]
// @Security     BearerAuth
func (s *Server) getBookingByIDHandler(c *gin.Context) {
	bookingID := c.Param("booking_id")
	c.JSON(http.StatusOK, gin.H{
		"booking_id":  bookingID,
		"venue_id":    "123e4567-e89b-12d3-a456-426614174000",
		"venue_name":  "Grand Palace",
		"user_id":     "789e4567-e89b-12d3-a456-426614174000",
		"status":      "CONFIRMED",
		"start_time":  "2026-06-25T10:00:00Z",
		"end_time":    "2026-06-25T14:00:00Z",
		"total_price": 600.00,
		"notes":       "Need AV setup and projector.",
		"created_at":  "2026-06-19T20:44:12Z",
	})
}

// cancelBookingHandler cancels an active booking
// @Summary      Cancel booking
// @Description  Cancel an active booking by its ID
// @Tags         bookings
// @Accept       json
// @Produce      json
// @Param        booking_id path string true "Booking ID"
// @Success      200  {object}  map[string]any
// @Router       /api/v1/bookings/{booking_id} [delete]
// @Security     BearerAuth
func (s *Server) cancelBookingHandler(c *gin.Context) {
	bookingID := c.Param("booking_id")
	c.JSON(http.StatusOK, gin.H{
		"booking_id": bookingID,
		"status":     "CANCELLED",
	})
}